* Fixed requirements.txt  * Refactor api.py  * Refactored URLs  * Refactored views, cleaned up imports, fixed bug with login  * Refactored main.html
