+++ flinder/flinderserver/urls.py
@@ -7,7 +7,7 @@ app_name = "flinder"
 
 urlpatterns = [
     path('', views.index, name="index"),
-    path('login/', views.login, name="login"),
+    path('login/', views.login_view, name="login"),
     path('register/', views.register, name="register"),
     path('register/account_type', views.register_account_type, name="register_account_type"),
     path('register/room_seeker', views.register_room_seeker, name="register_room_seeker"),
@@ -16,6 +16,6 @@ urlpatterns = [
     path('edit_profile', views.edit_profile, name="edit_profile"),
     path('main', views.main, name="main"),
     path('profile/<slug:profile_slug>', views.profile, name="profile"),
-    path('api/get_matches', views.get_matches, name="get_matches")
-    path('api/get_match', api.get_match, name="get_match")
+    path('api/get_cards', api.get_cards, name="get_cards"),
+    path('api/get_matches', api.get_matches, name="get_matches")
 ]
