+++ flinder/requirements.txt
@@ -1,0 +1,8 @@
+certifi==2021.10.8
+Django==2.2.27
+django-crispy-forms==1.14.0
+Pillow==9.0.1
+python-decouple==3.6
+pytz==2022.1
+sqlparse==0.4.2
+wincertstore==0.2
